Rename nested-loop-vs-hash to 1365-how-many-numbers-are-smaller-than-current-number.py
My solution to problem 1365. I have 2 solutions here, one with a nested for loop and one with hash code. The output of this particular set of code will output the time taken to run the program since that was my focus when solving this problem. I wanted to see exactly how much of a time difference it would be between a nested loop and a hash solution. Further directions are in the code.